Switch to an offsets-based representation
This representation is interesting because it allows for mixed cell type
topologies.

This is way more convenient than the previous implementation which took
the number of cell vertices as a const generic parameter, and I was not
sure how this would be generalized to different cell types in the same
mesh.