Refactor `locate_many` in functional style
I tried running some benchmarks to check if it's really faster that way,
the speedup was not very clear but at least it doesn't seem to be
slower. And it makes it a one-liner so let's keep it that way!

The strange thing is that I tried a for loop with `Vec::new` and another
one with `Vec::with_capacity` and the time difference was not making a
lot of sense. Whichever ran second was the faster one, even though in
principle the one that pre-allocates the right size directly should be
systematically faster.