Improve DCEL construction with `Option`s
Both vertices and half-edges have attributes that are not known when
the structs are created, so previously I used `usize::MAX` as a
placeholder, but this is not very nice.
I added partially initialized vertices and half-edges that have
`Option`s instead, and at the end these are converted to regular
vertices and half-edges.

This does allocate more memory because I need to allocate memory twice
to hold the two vectors, but this approach is cleaner IMO.

This made the `add_vertices` method useless, so I removed it.