Locate edge instead of point for insertion
We actually need to locate the whole edge instead of just the origin
because we need to compare slopes in the case of an edge whose origin
is already present in the trap map.

This lead to the implementation of the "Follow Segment" algorithm for
the book.

Also, add a couple of useful helper methods to `Dcel`.

Finally, fix the logic for mutating the `Dag` (some trapezoids were
wrong), and fix the order of the points in the bounding box (they were
clockwise).