Change `find_face` to `find_node`
We are actually looking for which node of the DAG corresponds to the
query point, because we want to be able to tell if the point lies on a
vertex or an edge of the mesh.

Then, the point locator can always return a face for points on the edges
and on the vertices.

This allows to get the correct behavior for the edge and corner cases in
the tests.