Make both Python and Rust libraries named `ploc`
The trick was to alias the Rust library in the Cargo.toml using the
syntax:
  <alias> = { package = "<true name>", path = "<path>" }

That way you can name the lib the same as the Rust crate, which means
the library can be named `ploc` on the Python side.

I also renamed the package in the PyO3 directory to `ploc-py` instead of
`py-ploc` for consistency with `ploc-rs` and to have them grouped
together in case of alphabetical sorting.