Switch to a single method on the Python side
There is now an optional argument to choose either the sequential or
parallel method for the computation. This enum also has an `auto`
variant in case the user would like for this choice to be made
automatically. This is not implemented yet but the idea would be to
choose one or the other depending on the size of the query (sequential
for small queries and parallel for large queries) but this requires
doing some benchmarks to evaluate the breakeven.