Use smallvec for child nodes in the DAG
In my application I know I always have 0 or 2 child nodes, so it's ideal
to use smallvec. This reduces indirection as the child nodes are now
stored as part of the `Node` struct instead of somewhere else on the
heap.

For very small meshes (~25 elements) this does not matter much, but for
larger meshes (tested with 2500 elements and more) I got a 20-30% speedup!

Also, the creation of the trapezoidal map is 20-30% faster.