Add edge endpoints to Y node
Only storing the half-edge id is more efficient in terms of memory
usage, but it comes at the cost of an extra layer of indirection when
traversing the DAG.
Every time we encounter a Y node, we get the half-edge id, then we need
to look up the hedge data in a heap-allocated Vec, which yields the
vertex ids of the endpoints whose coordinates have to be looked up in
another Vec.

It's more efficient to directly store the vertex ids of the endpoints
because this avoids one lookup.