Remove DCEL + fix list of edges
When checking whether an `Edge` was already in the hash map the result
was not right because of the presence of the face in the `Edge`.
Instead, I needed to just put the indices of the endpoints in the hash
map.

Additionally, the DCEL has been completely removed from the data
structure!