Make edge list deterministic
The edges are shuffled but the initial order was not deterministic
because of the iteration on the hash map of lefties.
I added the index where the lefties should be inserted in the list, as
well as made elements of the list `Option<Edge>`s, so that there are
slots available when we add the boundary edges.
There are some extra slots for lefties with a twin that are encountered
before their twin, but that seems to be acceptable.