Use NonMaxUsize to optimize `Option<usize>` size
A `usize` takes up 8 bytes, but an `Option<usize>` takes up 16 bytes.
This means that the places where we may have a missing `usize` take up
twice the amount of space.

Instead, using the `nonmax` crate allows to have optional integers that
take the same amount of space, at the price of not being able to have
the value `usize::MAX`. This is a small price to pay, as we are probably
going to have bigger problems if we have an array that's this big...

In the `regular_grid` example, this reduces the peak memory usage by
about 11%, which is already something.