Attempt at implementing the DAG with trait objects
This did reduce memory usage significantly, we got almost on par with
the mpl implementation in the 200x200 grid with under 80MB of peak
memory usage.

However the runtime performance is immensely degraded (at least 50%),
which was to be expected because this means that all the nodes are
allocated on the heap so there is a lot of indirection.

Ccl: not a good idea!