Push if outside of the for loop
This is basically a loop unrolling.

It was a bit of a shame that we were having this big if statement in
the loop, because the loop really only deals with the middle trapezoids.
For instance, when exactly one trapezoid is intersected, there is no
loop to do, which is not easily seen in the previous implementation.

This might lead to a small speedup in the construction of the trap map.