Fix bug for multiply connected geometries
We were only storing a single face per edge, namely the one we were
traversing when we met the edge.
As a consequence, trapezoids that were contained inside of a hole in
the domain had both top and bottom edges with an assigned face, while
it should have been `None`.

To fix it, we need to store for each edge the face above and the face
below, and they are both optional. That way, we can represent the
fact that an edge may have a face above but not below and vice versa.
This is actually what matplotlib does.

Another added benefit is that we can check the consistency of the
trap map by checking that the face above the bottom edge, and the face
below the top edge of a trapezoid are the same.