Perf: get rid of parents vec (#9)
We only used the `parents` field for a few sanity checks, but it's not
actually that useful.
The only thing is it made computing the depth of a node much more
efficient, but again since that's something only used for debugging it's
fine if it's slower now.
Since we get smaller memory usage from removing this field it's
definitely worth it.